Refactor PriorityQueue
Fix an issue with the Insert Function. The CapacityDepth was not updated properly. This caused the Store to be bigger than needed.

Fix an issue with the PercolateUp Function. The parent was compared to the child, rather than the data stored at index 0.

Rename some properties and functions to better reflect their usage. Namely these are the Depth, now named CapacityDepth as the depth the the depth of the capacity, not the data, and the DeleteMin function, now named Dequeue. DeleteMin is renamed to have the PriorityQueue look more like the built-in Queue. Add several comments as well as some extra unit tests.