Fix issue #1
Add several more unit tests for Indexer, Find and Union.

Refactor Find to no longer be recursive for better support of large datasets. This refactor also solved the problem with Find that existed before.

Add a check to Union to ensure that the to be united roots are not the same. This caused OutOfMemory exceptions.