Add IVector and Vector
Add the IVector interface for all vectors to use. This interface supports
several arithmetic functions as well as other useful vector functions. It
also has a factory methods to create more vectors of the same type.

Add the Vector class. This is a relatively generic implementation with an
array as store, allowing for a variable sizes.